+++ src/private/release/PublishNovaBuiltModule.ps1
@@ -1,5 +1,5 @@
 function Publish-NovaBuiltModule {
-    [CmdletBinding(DefaultParameterSetName = 'Local')]
+    [CmdletBinding(DefaultParameterSetName = 'Local', SupportsShouldProcess = $true)]
     param(
         [Parameter(ParameterSetName = 'Repository')]
         [string]$Repository,
@@ -14,7 +14,12 @@ function Publish-NovaBuiltModule {
     }
 
     if ( $PSBoundParameters.ContainsKey('Repository')) {
-        Publish-NovaBuiltModuleToRepository -ProjectInfo $ProjectInfo -Repository $Repository -ApiKey $ApiKey
+        $shouldRun = $PSCmdlet.ShouldProcess($Repository, 'Publish built module to repository')
+        if (-not $shouldRun -and -not $WhatIfPreference) {
+            return
+        }
+
+        Publish-NovaBuiltModuleToRepository -ProjectInfo $ProjectInfo -Repository $Repository -ApiKey $ApiKey -WhatIf:$WhatIfPreference -Confirm:$false
         return
     }
 
@@ -23,5 +28,10 @@ function Publish-NovaBuiltModule {
         $resolvedModuleDirectoryPath = Resolve-NovaLocalPublishPath -ModuleDirectoryPath $ModuleDirectoryPath
     }
 
-    Publish-NovaBuiltModuleToDirectory -ProjectInfo $ProjectInfo -ModuleDirectoryPath $resolvedModuleDirectoryPath
+    $shouldRun = $PSCmdlet.ShouldProcess($resolvedModuleDirectoryPath, 'Publish built module to local directory')
+    if (-not $shouldRun -and -not $WhatIfPreference) {
+        return
+    }
+
+    Publish-NovaBuiltModuleToDirectory -ProjectInfo $ProjectInfo -ModuleDirectoryPath $resolvedModuleDirectoryPath -WhatIf:$WhatIfPreference -Confirm:$false
 }