+++ src/private/shared/GetNovaShouldProcessForwardingParameter.ps1
@@ -1,0 +1,14 @@
+function Get-NovaShouldProcessForwardingParameter {
+    [CmdletBinding()]
+    param(
+        [switch]$WhatIfEnabled
+    )
+
+    $forwardingParameter = @{Confirm = $false}
+    if ($WhatIfEnabled) {
+        $forwardingParameter.WhatIf = $true
+    }
+
+    return $forwardingParameter
+}
+