+++ src/public/InvokeNovaBuild.ps1
@@ -1,11 +1,16 @@
 function Invoke-NovaBuild {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess = $true)]
     param (
     )
-    Reset-ProjectDist
-    Build-Module
 
     $data = Get-NovaProjectInfo
+    if (-not $PSCmdlet.ShouldProcess($data.OutputModuleDir, 'Build Nova module output')) {
+        return
+    }
+
+    Reset-ProjectDist -Confirm:$false
+    Build-Module
+
     if ($data.FailOnDuplicateFunctionNames) {
         Assert-BuiltModuleHasNoDuplicateFunctionName -ProjectInfo $data
     }